\name{as.ggraph}
\alias{as.ggraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert a qgraph object to ggraph object
}
\description{
This is an experimental function to convert a qgraph object to ggraph object
}
\usage{
as.ggraph(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A \code{qgraph} object
}
}

\author{
Sacha Epskamp <mail@sachaepskamp.com>
}

