\name{big5}
\alias{big5}
\docType{data}
\title{
Big 5 dataset
}
\description{
This is a dataset of the Big 5 personality traits that will be used in talks and the paper.
}
\usage{data(big5)}
\format{
  The format is:
 num [1:500, 1:240] 2 3 4 4 5 2 2 1 4 2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:240] "N1" "E2" "O3" "A4" ...
}
