\name{big5groups}
\alias{big5groups}
\docType{data}
\title{
Big 5 groups list
}
\description{
This is the groups list of the big 5 data
}
\usage{data(big5groups)}
\format{
  The format is:
List of 5
 $ Neuroticism      : num [1:48] 1 6 11 16 21 26 31 36 41 46 ...
 $ Extraversion     : num [1:48] 2 7 12 17 22 27 32 37 42 47 ...
 $ Openness         : num [1:48] 3 8 13 18 23 28 33 38 43 48 ...
 $ Agreeableness    : num [1:48] 4 9 14 19 24 29 34 39 44 49 ...
 $ Conscientiousness: num [1:48] 5 10 15 20 25 30 35 40 45 50 ...
}
