\name{print.qgraph}
\alias{print.qgraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print edgelist
}
\description{
This function prints the edgelist of a "qgraph" object
}
\usage{
\method{print}{qgraph}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A "qgraph" object
}
\item{...}{
These arguments are not used
}
}

\author{
Sacha Epskamp (qgraph@sachaepskamp.com)
}

\seealso{
\code{\link{qgraph}}
}
