\name{qgraph-package}
\alias{qgraph-package}
\docType{package}
\title{
qgraph
}
\description{
This package can be used to visualize psychometric data. See \code{\link{qgraph}} for the main function.}
\details{
\tabular{ll}{
Package: \tab qgraph\cr
Type: \tab Package\cr
Version: \tab 1.2.5\cr
Date: \tab 2014-05-22\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sacha Epskamp (mail@sachaepskamp.com)

Maintainer: Sacha Epskamp<mail@sachaepskamp.com>
}
\references{
Sacha Epskamp, Angelique O. J. Cramer, Lourens J.
  Waldorp, Verena D. Schmittmann, Denny Borsboom (2012).
  qgraph: Network Visualizations of Relationships in
  Psychometric Data. Journal of Statistical Software,
  48(4), 1-18. URL http://www.jstatsoft.org/v48/i04/.

}

\keyword{ package }


