\name{summary.qgraph}
\alias{summary.qgraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary method for "qgraph"
}
\description{
This function creates a brief summary based on a "qgraph" object.
}
\usage{
\method{summary}{qgraph}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A "qgraph" object
}
\item{...}{
These arguments are not used
}
}

\author{
Sacha Epskamp (mail@sachaepskamp.com)
}

\seealso{
\code{\link{qgraph}}
}
