\encoding{UTF8}
\name{big5groups}
\alias{big5groups}
\docType{data}
\title{
Big 5 groups list
}
\description{
This is the groups list of the big 5 data. It is a measurement of the Dutch translation of the NEO-PI-R on 500 first year psychology students (Dolan, Oort, Stoel, Wicherts, 2009).
}
\usage{data(big5groups)}
\references{
Hoekstra, H. A., Ormel, J., & De Fruyt, F. (2003). NEO-PI-R/NEO-FFI: Big 5 persoonlijkheidsvragenlijst. Handleiding [Manual of the Dutch version of the NEO-PI-R/NEO-FFI]. Lisse, The Netherlands: Swets and Zeitlinger.

Dolan, C. V., Oort, F. J., Stoel, R. D., & Wicherts, J. M. (2009). Testing measurement invariance in the target rotates multigroup exploratory factor model. Structural Equation Modeling, 16, 295–314.
}
\format{
  The format is:
List of 5
 $ Neuroticism      : num [1:48] 1 6 11 16 21 26 31 36 41 46 ...
 $ Extraversion     : num [1:48] 2 7 12 17 22 27 32 37 42 47 ...
 $ Openness         : num [1:48] 3 8 13 18 23 28 33 38 43 48 ...
 $ Agreeableness    : num [1:48] 4 9 14 19 24 29 34 39 44 49 ...
 $ Conscientiousness: num [1:48] 5 10 15 20 25 30 35 40 45 50 ...
}
