\name{plot.qgraph}
\alias{plot.qgraph}

\title{
Plot method for "qgraph"
}
\description{
This function is exactly the same as \code{\link{qgraph}}
}
\usage{
\method{plot}{qgraph}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A "qgraph" object
}
  \item{\dots}{
Arguments send to \code{\link{qgraph}}
}
}

\author{
Sacha Epskamp (mail@sachaepskamp.com)
}
