\name{qgraph-package}
\alias{qgraph-package}
\docType{package}
\title{
qgraph
}
\description{
This package can be used to visualize psychometric data. See \code{\link{qgraph}} for the main function.}
\details{
\tabular{ll}{
Package: \tab qgraph\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sacha Epskamp

Maintainer: Sacha Epskamp<s.epskamp@uva.nl>
}
\references{
https://sites.google.com/site/qgraphproject
}

\keyword{ package }


